backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
pydantic==2.7.4
fastapi-cors==0.0.1
python-dotenv
